+++ src/gui/widgets/dialogs/gw_impl_ui.cpp
@@ -14,6 +14,7 @@
 #include "model/shapes/image.hpp"
 #include "model/shapes/repeater.hpp"
 #include "model/shapes/trim.hpp"
+#include "model/shapes/inflate_deflate.hpp"
 
 #include "widgets/dialogs/io_status_dialog.hpp"
 #include "widgets/dialogs/about_dialog.hpp"
@@ -194,6 +195,7 @@ void GlaxnimateWindow::Private::init_actions()
     });
     add_modifier_menu_action<model::Repeater>(ui.menu_new_layer);
     add_modifier_menu_action<model::Trim>(ui.menu_new_layer);
+    add_modifier_menu_action<model::InflateDeflate>(ui.menu_new_layer);
     connect(ui.action_import_image, &QAction::triggered, parent, [this]{import_image();});
     connect(ui.action_delete, &QAction::triggered, parent, &GlaxnimateWindow::delete_selected);
     connect(ui.action_export, &QAction::triggered, parent, &GlaxnimateWindow::document_export);