+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 * Editing:
     * Action to make animations loop in the property context menu
+    * Added Inflate/Deflate modifier (aka Pucker/Bloat)
 * I/O:
     * Loading Lottie with `meta` no longer show a warning
     * Fixed UUIDs when loading lottie animations